mshv: Simplify vCPU event loop
Effectively, vCPU thread is waiting for a message from hypervisor and this
patch reflects it in the code.

In addition it makes the waiting killable (not interruptible) for
now as it would allow to expose the logic simplification.
A feature to interrupt vCPU wait by a signal will be introduced in another patch.

This patch brings back simultaneous clearing of both intercept and explicit
suspend, but in the opposite order: the intercept one is cleared first, which
guarantees there won't be racing with the message delivery.

Signed-off-by: Stanislav Kinsburskiy <stanislav.kinsburskiy@gmail.com>