makefile: Fix EXTRAVERSION
As arranged with the Mariner team, the version used in RPMs and tarballs
has to look like `mshv1`. The version is also to be changed anytime code
changes. For this to work, the EXTRAVERSION in the Makefile needs to be
updated. From there, it'll be also reflected in config and later during
the `make install` and other targets. As the rolling branch is a mix of
LSG-linux-rolling and Dom0, and is also used for releases, the versions
need to be adapted cleanly at any points where code would go into tag.

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>