pipelines: run checkpatch.pl on pull requests
Add a script to run checkpatch.pl on the PR commits and post the issues
found as comments on the PR. Fetching the commits in a PR and posting a
comment is done via the Azure DevOps REST API.

The checkpatch step in the pipeline should be optional because there
could be cases where we would want to ignore the warnings issued by
checkpatch. As per the upstream kernel docs [1]:

"Checkpatch is not always right. Your judgement takes precedence over
checkpatch messages. If your code looks better with the violations, then
its probably best left alone."

[1]: https://docs.kernel.org/dev-tools/checkpatch.html

Signed-off-by: Anirudh Rayabharam <anrayabh@linux.microsoft.com>