hv: reboot: Implement EnterSleepState hypecrall
The hypercall is issued in the callabck attached to the reboot syscall
notification. The reboot notifier appears to be equivalent to those
provided for ACPI, however it appears to be more reliable as it'll be
triggered also when ACPI support is not present.

Internally, the EnterSleepState hypercall needs to have a sleep state
prepared in the hypervisor. The sleep state initialization is done
through another call HvSetSystemProperty. This patch adds the
rudimentary parts necessary for this hypercall to succeed setting sleep
states. More system properties are available to set and thus the scructs
and use cases can be extended as the development is going forward.

Fixes: #40833429

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>