mshv: Use dedicated variable to kick VP
This looks like a minor cleanup change, but it's actually a workaround for
some nasty hardware-related (not cause) bug, leading to soft lockup in L2 guest.

Clearing intercept_suspend bit right before waiting for an event, represented
by a sibling kicked_by_hv bit, is not atomic and thus sometimes leads to the
latter being dropped as well, which in turn leads to VP thread getting stuck
in waiting for the bit, set by the message handler.

Signed-off-by: Stanislav Kinsburskiy <stanislav.kinsburskiy@gmail.com>