Drivers: hv: Add dummy handler for async hypercall intercept
Some root hypercalls may return a status of `HV_STATUS_CALL_PENDING`
to indicate that the hypercall will complete asynchronously.

When it completes MSHV returns an interrupt marking as completion
of the hypercall. Thus, we need to install an intercept handler to
handle the interrupt.

Currently it is just a dummy handler and will add more details in the
future commits.

Signed-off-by: Jinank Jain <jinankjain@linux.microsoft.com>