pipeline: Implement tooling to pull upstream in pipeline
The tools implement the base for the upstream tag pulling automation.
The corresponding pipeline is planned to be triggered as a cronjob and
will create a feature branch and a PR with the updates pulled from the
upstream.
The pipeline can also be triggered manually.

This is supposed to work most of the time. It is however not excluded,
that sometimes there will be merge conflicts that cannot be resolved. In
that case a manual intrusion will be needed or possibly the auto merge
algorithm can be improved, too.

Fixes: #42529873

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>