pipelines: Pull the checkpatch.pl invocation out from the matrix
The patch set is always the same, so the task to be repeated at every
build doesn't make much effect. Instead, it is moved into a separate
job. The additional overhead of checking out and setting up an agent is
negligible. Also, this job is a good candidate to be later extended with
more code validation tasks like static analysis.

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>