pipeline: Create PR validation pipeline
These are the changes to create PR pipeline for linux-dom0
pull request. This will run some default tests when the PR
is raised. We can run it manually as well for given branch
having flexibilty to select the testcases while running
the pipeline as per the need.

This will invoke the linux dom0 template from lsg-lisa
repo and run the testcases as per the selection. Default
tests we will run is:

Functional: smoke_test, l3_cache_check, cpu_count_check
      MSHV: mshv stress test, rust vmm test

Signed-off-by: Smit Gardhariya <sgardhariya@microsoft.com>