Merged PR 7443383: Enable MAXSMP
In the Linux kernel configuration, an easy option to set is MAXSMP​ to allow up to 8192​ processors.
The size of the kernel increases by ~0.5MiB, and I have not noticed an increased runtime consumption with my naked eye and the free​ utility.
The perf team could scrutinize that of course.

The 2048​ processors require updating the kernel configuration logic well outside of the code the company touches, looks very hacky. I don't want to embark on adding an option for 2048​ because of that. Maybe we could use 8192​ for now and ask LSG to make it 2048​ in the future, applying their expertise.

Fixes:
* https://microsoft.visualstudio.com/OS/_workitems/edit/40477660 (Underhill blocker?)
* https://microsoft.visualstudio.com/OS/_workitems/edit/42865254 (hyp branch RI blocker?)

240 processors in an Underhill VM:
![image.png](https://msazure.visualstudio.com/6e09f0eb-e532-4615-98ab-afa7a6220e98/_apis/git/repositories/82197ee4-cbe1-4fed-bb27-d2d74d9fc22d/pullRequests/7443383/attachments/image.png)