pipeline: checkpatch: Introduce ignore list
In this particular case, the warning is thrown when pulling an upstream
tag and the commit message contains some unreachable commit hash. To
make this work properly, it probably needs to pull ALL the history from
the remote including all the tags and branches. This would take time
that is most likely not to be spent. Especially pulls from upstream
might contain dozens of such commits mentioning other commit hashes that
are unreachable, thus enforcing to click multiple dozens of times on the
PR page to resolve.

In other cases, where commit hashes are put with a typo or are wrong,
the errors won't be caught as well.
Given our default checkout strategy is to only have the merge branch,
that will most likely be not relevant as even then any hash mention
would not pass the validation.

An example error message shown:

```
WARNING: Unknown commit id '46a525e199e4037516f7e498c18f065b09df32ac',
maybe rebased or not pulled?
commit 46a525e199e4037516f7e498c18f065b09df32ac upstream.
```

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>