azuredevops: Add template section for recording changelog
The content between "Changelog record" and "References" can be
automatically parsed out when time comes to compose the changelog.
Perhaps a stronger format is needed but starting as simple as this will
give the base for implementing some tooling to gather all the records.
The format and the tools can be adapted in the future to match the
needs.

Fixes: #41983415

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>