mshv: Assert virtual interrupt asynchronously
Asserting the virtual interrupt during eventfd_write is a bottleneck for workloads
with high I/O operations.
Offloading virtual interrupt injection to a kernel worker allows to improve
bandwidth for block IO by 8% and 12% for write and read IO respectively as well
as for network IO by 4% and 6% for network rx and tx respectively.

Signed-off-by: Stanislav Kinsburskiy <stanislav.kinsburskiy@gmail.com>