Merged PR 7610405: No warning when a host doesn't provide reg page in VSM
Don't warn the user if the reg page is not supported.

That is a supported condition. No warnings should be issued if the hypervisor doesn't support register page as this is expected for older hypervisors. The reg page is a performance optimization, not a hard requirement.

NOTE the TODO in the code.

Signed-Off-By Roman Kiselev <romank@microsoft.com>