Hack: Call vtl1 init in the boot sequence
Instead of having a standard init call for vtl1 init, call it
in the boot sequence to return back to vtl0 prior to scheduler is
kick started and the rootfs is loaded.

This is a temporary hack till we manage to load a initramfs to unblock
rest of the team.

Signed-off-by: Thara Gopinath <tgopinath@microsoft.com>