Merged PR 48: Add initrd and register ioctl to return to VTL0
This PR requires the **initrd.cpio** file, which is generated by https://dev.azure.com/lsg-linux-core/Linux-VBS/_git/sk-tiny-initramfs, **to be present in the repo directory**.

# Summary of the PR

Describe what issue this PR solves. Why it is needed.

# How the issue is solved?

Describe how the issue is solved.

# How is this PR tested?

What are the tests done?

# Changelog record

Anything relevant for the changelog, otherwise N/A.

# References

Provide any references that may help the reviewers (documents, code etc)

*Please make sure patches are in "upstream-ready" state. Please check out this [link](https://www.kernel.org/doc/html/latest/process/submitting-patches.html) for guidance.*

Add initrd and register ioctl to return to VTL0