Merged PR 79: Create per cpu kthread for handling VTL1 loop
Create per cpu kthread for handling VTL1 loop

VTL1 entry exit runs in a loop. Create a per-cpu kthread on which this loop can run and pin it to each CPU.
Also start the kthread for boot cpu on return to vtl0 ioctl call.

For now put the init process to sleep in kernel.

Signed-off-by: Thara Gopinath <tgopinath@microsoft.com>