Boot secondary VPs
Adds up to 63 Virtual Application/Secondary Processors to VTL1.

This includes 2 vtlcalls:
- `VSM_VTL_CALL_FUNC_ID_ENABLE_APS_VTL`: This function will use VTL0's
  cpu_present_mask to make CPU's present in VTL1. It will also enable VTL1
for each CPU.
- `VSM_VTL_CALL_FUNC_ID_BOOT_APS`: This function will use VTL0's
  cpu_online_mask and will boot each CPU that is set.

A key part of bringing up Application Processor `x` (APx)  is that it is
done in part by the Boot CPU and in part by APx. The Boot CPU in VTL1
initiates the Boot Process of APx and at some point it needs to notify APx
in VTL0 that it can now transition to VTL1 to continue its part of the boot
process. This is done by using the `boot_signal` shared page between VTL0
and VTL1.

Signed-off-by: Anna Trikalinou <atrikalinou@microsoft.com>
Signed-off-by: Angelina Vu <angelinavu@microsoft.com>
Signed-off-by: Thara Gopinath <tgopinath@microsoft.com>