Add support to set EPT permissions for VTL0 memory
Add a new service id that enables a guest to set EPT permissions
for guest pages.

Until now, all of the guest pages are given RWX permissions in the EPT.
In Heki, we want to restrict the permissions to what is strictly needed.
For instance, a text page only needs R_X. A read-only data page only
needs R__. A normal data page only needs RW_.

The guest will pass a page list to the hypercall. The page list is a
list of one or more physical pages each of which contains a array of
guest ranges and attributes. Currently, the attributes only contain
permissions. In the future, other attributes may be added.  The
hypervisor will apply the specified permissions in the EPT.

When a guest try to access its memory in a way which is not allowed, the
hypervisor sends a fault to the guest.

Co-developed-by: Angelina Vu <angelinavu@microsoft.com>
Signed-off-by: Angelina Vu <angelinavu@microsoft.com>
Signed-off-by: Madhavan T. Venkataraman <madvenka@linux.microsoft.com>