heki: x86: Protect guest kernel memory
Pass guest kernel pages along with their permissions to the host so
that EPT permissions can be set for them.

This is done at the end of kernel init just before kicking off the
init process. Up to this point, the kernel is trusted.

Beyond this point, any request to modify EPT permissions for guest
pages needs to be authenticated in the host. E.g., module loading and
unloading, text patching, eBPF JIT.

This support will be implemented in the future. Until then, modules
have to be built in and the other features have to be disabled.

Co-developed-by: Mickaël Salaün <mic@digikod.net>
Signed-off-by: Mickaël Salaün <mic@digikod.net>
Signed-off-by: Madhavan T. Venkataraman <madvenka@linux.microsoft.com>