Send kernel data to VTL1
Kernel data needs to be sent from VTL0 to VTL1 for various purposes.
The VTL call load_kdata() is implemented for this.

For starters, send module certificates. VTL1 will create a trusted
keyring from the module certificates.

In the future, this will be used by VTL1 to verify the signature of a
module when the module is loaded in VTL0.

A kernel data structure is virtually contiguous, but not necessarily
physically contiguous. So, multiple ranges may be sent for a single
data structure. Also, multiple data structures can be sent in a single
call.

To manage this, assign a unique identifier to each kernel data structure.
Set the attributes field of each range in a data structure to the
identifier. This enables VTL0 to gather the ranges of a single data
structure and map them in a single mapping.

Signed-off-by: Madhavan T. Venkataraman <madvenka@linux.microsoft.com>