Receive a VTL0 module and validate its contents
Implement a hypercall to receive a module blob from VTL0 and
check the module's signature against the trusted keyring in VTL1.

After the signature is checked successfully, execute the steps to layout
and allocate the module sections in VTL1 just like in VTL0.

Compare the module sections constructed in VTL1 with the ones passed
from VTL0. If they match, then the guest module contents are validated.

Return an authentication token to VTL0 so VTL0 can refer to VTL1's copy
of the module in the future.

Signed-off-by: Madhavan T. Venkataraman <madvenka@linux.microsoft.com>