Apply post relocation fixes to VTL0 module in VTL1
Many post relocation fixups happen for modules. At this time, HEKI
does not support all of them. Unconfigure the ones that are not
supported. The following are supported:

	- Paravirt fixups
	- CPU alternatives
	- SMP locks

Perform these fixups for module contents independently in VTL1.

Finally, compare the contents with the VTL0 module contents passed. If
they match, then the VTL0 module contents are validated.

Signed-off-by: Madhavan T. Venkataraman <madvenka@linux.microsoft.com>