Merged PR 159: Runtime Secondary Key Support Added
When keys are added to the secondary keyring, send the key payload data to vtl1 where it is verified, and corresponding key is added to the vtl0 trusted keyring.

Signed-off-by: Daniel Urbonas t-durbonas@microsoft.com

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Runtime support for secondary key data.

#### PR Summary
This pull request introduces runtime support for copying secondary key data to the hypervisor.
- `virt/heki/module.c`: Added `heki_copy_secondary_add_key_data` function to handle copying secondary key data.
- `drivers/hv/hv_vsm.c`: Implemented `hv_vsm_copy_secondary_key_data` function and updated `hyperv_heki_hypervisor` structure.
- `security/keys/keyctl.c`: Integrated secondary key data copying in the `add_key` syscall.
- `include/linux/heki.h`: Added function prototype for copying secondary key data.
- `drivers/hv/hv_vsm.h`: Defined VSM call function ID for copying secondary key data.